requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
billiard==4.2.0
boto3==1.35.4
botocore==1.35.4
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
distro==1.9.0
Django==5.0.6
et-xmlfile==1.1.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.7
jiter==0.8.2
jmespath==1.0.1
kombu==5.4.1
logfury==1.0.1
mailersend==0.5.8
Markdown==3.6
openai==1.61.1
openpyxl==3.1.5
packaging==24.1
pillow==10.3.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
redis==5.0.8
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
